// apps/admin/src/generated/api/schemas/listFilesParams.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { ListFilesStorageDriver } from "./listFilesStorageDriver";
import type { Object } from "./object";

export type ListFilesParams = {
  /**
   * @minimum 1
   */
  page?: Object;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: Object;
  search?: string;
  /**
   * Sort by one of: displayName, mimeType, size, storageDriver, createdAt, updatedAt
   */
  sort?: string;
  mimeType?: string;
  storageDriver?: ListFilesStorageDriver;
};